requirements.txt
discord.py
wavelink~=3.1.2
lavalink
aiohttp~=3.9.1
python-dotenv~=1.0.0
craiyon.py


